refactor redondant code
