+++ src/components/Dashboard/ChangingPlans.component.js
@@ -15,6 +15,7 @@ import {
 	getCheckoutURL,
 	getCustomerPortalLink,
 } from '@/services/lemonsqueezy.service'
+import { sortAndSetPlans } from '@/utils/sortAndSetPlans'
 
 const frequencies = [
 	{ value: 'monthly', label: 'Monthly', priceSuffix: '/month' },
@@ -51,20 +52,7 @@ export function ChangingPlansComponent() {
 
 	useEffect(() => {
 		getPlans().then(plans => {
-			plans.sort((a, b) => {
-				// Extract the numbers from the plan names
-				const numA = parseInt(a.name.match(/\d+/)?.[0] || '0')
-				const numB = parseInt(b.name.match(/\d+/)?.[0] || '0')
-
-				// Compare the numbers
-				if (numA !== numB) {
-					return numA - numB
-				}
-
-				// If the numbers are the same, compare the names
-				return b.name.localeCompare(a.name)
-			})
-			setPlans(plans)
+			sortAndSetPlans(plans, setPlans)
 		})
 		checkSubscription()
 	}, [])
