+++ src/components/Dashboard/RefillPlans.component.js
@@ -12,6 +12,7 @@ import {
 	getCheckoutURL,
 	getCustomerPortalLink,
 } from '@/services/lemonsqueezy.service'
+import { sortAndSetPlans } from '@/utils/sortAndSetPlans'
 
 function classNames(...classes) {
 	return classes.filter(Boolean).join(' ')
@@ -27,20 +28,7 @@ export function RefillPlansComponent() {
 
 	useEffect(() => {
 		getPlans().then(plans => {
-			plans.sort((a, b) => {
-				// Extract the numbers from the plan names
-				const numA = parseInt(a.name.match(/\d+/)?.[0] || '0')
-				const numB = parseInt(b.name.match(/\d+/)?.[0] || '0')
-
-				// Compare the numbers
-				if (numA !== numB) {
-					return numA - numB
-				}
-
-				// If the numbers are the same, compare the names
-				return b.name.localeCompare(a.name)
-			})
-			setPlans(plans)
+			sortAndSetPlans(plans, setPlans)
 		})
 		checkSubscription()
 	}, [])
