+++ src/components/Landing/Pricing/Pricing.component.js
@@ -6,6 +6,7 @@ import { useEffect, useState } from 'react'
 
 import { SkeletonLoaderPricing } from '@/components/Skeletons/SkeletonLoaderPricing'
 import { getPlans } from '@/services/database.service'
+import { sortAndSetPlans } from '@/utils/sortAndSetPlans'
 
 const frequencies = [
 	{ value: 'monthly', label: 'Monthly', priceSuffix: '/month' },
@@ -37,20 +38,7 @@ export function PricingComponent() {
 
 	useEffect(() => {
 		getPlans().then(plans => {
-			plans.sort((a, b) => {
-				// Extract the numbers from the plan names
-				const numA = parseInt(a.name.match(/\d+/)?.[0] || '0')
-				const numB = parseInt(b.name.match(/\d+/)?.[0] || '0')
-
-				// Compare the numbers
-				if (numA !== numB) {
-					return numA - numB
-				}
-
-				// If the numbers are the same, compare the names
-				return b.name.localeCompare(a.name)
-			})
-			setPlans(plans)
+			sortAndSetPlans(plans, setPlans)
 		})
 	}, [])
